plugins/ocp: Refactor to clear feature identifier C1h and C3h
